User profile image upload
Added the image upload option on users profile ( show method ) so that the user is able to update publicly visible information. We check if the image isset in the request and if we validate the size,mime,type give it name, upload it move to public folder and save it in the db. If not just update other information